+++ test/shpec/operation_shpec.sh
@@ -181,7 +181,7 @@ function test_basic_ssh_operations ()
 			password="$(
 				docker logs \
 					ssh.pool-1.1.1 \
-				| awk '/^password :.*$/ { print $3 }'
+				| awk '/^password :.*$/ { print $3; }'
 			)"
 
 			it "Can be retrieved from the log."
@@ -438,14 +438,14 @@ function test_custom_ssh_configuration ()
 				password="$(
 					docker logs \
 						ssh.pool-1.1.1 \
-					| awk '/^password :.*$/ { print $3 }'
+					| awk '/^password :.*$/ { print $3; }'
 				)"
 
 				pwd_result="$(
 					expect test/ssh-password-auth.exp \
 						"${password}" \
 						app-admin \
-						127.0.0.1 \
+						"${DOCKER_HOSTNAME}" \
 						"${container_port_22}" \
 						"pwd" \
 					| tail -n 1 \
@@ -475,7 +475,8 @@ function test_custom_ssh_configuration ()
 
 			it "Logs the setting value."
 				password_authentication="$(
-					docker logs ssh.pool-1.1.1 \
+					docker logs \
+						ssh.pool-1.1.1 \
 					| awk '/^password authentication :.*$/ { print $0; }'
 				)"
 
@@ -583,7 +584,8 @@ function test_custom_ssh_configuration ()
 
 			it "Logs the setting value."
 				user_sudo="$(
-					docker logs ssh.pool-1.1.1 \
+					docker logs \
+						ssh.pool-1.1.1 \
 					| awk '/^sudo :.*$/ { print $0; }'
 				)"
 
@@ -644,7 +646,8 @@ function test_custom_ssh_configuration ()
 
 			it "Logs the setting value."
 				user="$(
-					docker logs ssh.pool-1.1.1 \
+					docker logs \
+						ssh.pool-1.1.1 \
 					| awk '/^user :.*$/ { print $0; }'
 				)"
 
@@ -705,7 +708,8 @@ function test_custom_ssh_configuration ()
 
 			it "Logs the key signature."
 				user_key_signature="$(
-					docker logs ssh.pool-1.1.1 \
+					docker logs \
+						ssh.pool-1.1.1 \
 					| awk '/^45:46:b0:ef:a5:e3:c9:6f:1e:66:94:ba:e1:fd:df:65$/ { print $1; }'
 				)"
 
@@ -796,14 +800,16 @@ function test_custom_ssh_configuration ()
 
 				it "Logs the key signatures."
 					user_key_signature="$(
-						docker logs ssh.pool-1.1.1 \
+						docker logs \
+							ssh.pool-1.1.1 \
 						| awk '/^45:46:b0:ef:a5:e3:c9:6f:1e:66:94:ba:e1:fd:df:65$/ { print $1; }'
 					)"
 
 					user_key_signature+=" "
 
 					user_key_signature+="$(
-						docker logs ssh.pool-1.1.1 \
+						docker logs \
+							ssh.pool-1.1.1 \
 						| awk '/^b3:2e:5d:8c:76:d3:c7:24:13:a3:4f:6f:4d:a2:31:9c$/ { print $1; }'
 					)"
 
@@ -879,14 +885,16 @@ function test_custom_ssh_configuration ()
 
 				it "Logs the key signatures."
 					user_key_signature="$(
-						docker logs ssh.pool-1.1.1 \
+						docker logs \
+							ssh.pool-1.1.1 \
 						| awk '/^45:46:b0:ef:a5:e3:c9:6f:1e:66:94:ba:e1:fd:df:65$/ { print $1; }'
 					)"
 
 					user_key_signature+=" "
 
 					user_key_signature+="$(
-						docker logs ssh.pool-1.1.1 \
+						docker logs \
+							ssh.pool-1.1.1 \
 						| awk '/^b3:2e:5d:8c:76:d3:c7:24:13:a3:4f:6f:4d:a2:31:9c$/ { print $1; }'
 					)"
 
@@ -944,7 +952,8 @@ function test_custom_ssh_configuration ()
 
 				it "Logs the key signature."
 					user_key_signature="$(
-						docker logs ssh.pool-1.1.1 \
+						docker logs \
+							ssh.pool-1.1.1 \
 						| sed -n -e '/^rsa private key fingerprint :$/{ n; p; }' \
 						| awk '{ print $1; }'
 					)"
@@ -1002,7 +1011,8 @@ function test_custom_ssh_configuration ()
 
 				it "Logs the key signature."
 					user_key_signature="$(
-						docker logs ssh.pool-1.1.1 \
+						docker logs \
+							ssh.pool-1.1.1 \
 						| sed -n -e '/^rsa private key fingerprint :$/{ n; p; }' \
 						| awk '{ print $1; }'
 					)"
@@ -1066,7 +1076,8 @@ function test_custom_ssh_configuration ()
 
 			it "Logs the setting value."
 				user_home="$(
-					docker logs ssh.pool-1.1.1 \
+					docker logs \
+						ssh.pool-1.1.1 \
 					| awk '/^home :.*$/ { print $0; }'
 				)"
 
@@ -1128,7 +1139,8 @@ function test_custom_ssh_configuration ()
 
 			it "Logs the setting value."
 				user_id="$(
-					docker logs ssh.pool-1.1.1 \
+					docker logs \
+						ssh.pool-1.1.1 \
 					| awk '/^id :.*$/ { print $0; }'
 				)"
 
@@ -1190,7 +1202,8 @@ function test_custom_ssh_configuration ()
 
 			it "Logs the setting value."
 				user_id="$(
-					docker logs ssh.pool-1.1.1 \
+					docker logs \
+						ssh.pool-1.1.1 \
 					| awk '/^id :.*$/ { print $0; }'
 				)"
 
@@ -1253,7 +1266,8 @@ function test_custom_ssh_configuration ()
 
 			it "Logs the setting value."
 				user_shell="$(
-					docker logs ssh.pool-1.1.1 \
+					docker logs \
+						ssh.pool-1.1.1 \
 					| awk '/^shell :.*$/ { print $0; }'
 				)"
 
@@ -1369,7 +1383,8 @@ function test_custom_ssh_configuration ()
 
 			it "Logs a redacted value."
 				user_password="$(
-					docker logs ssh.pool-1.1.1 \
+					docker logs \
+						ssh.pool-1.1.1 \
 					| awk '/^password :.*$/ { print $0; }'
 				)"
 
@@ -1440,7 +1455,8 @@ function test_custom_ssh_configuration ()
 
 			it "Logs a redacted value."
 				user_password="$(
-					docker logs ssh.pool-1.1.1 \
+					docker logs \
+						ssh.pool-1.1.1 \
 					| awk '/^password :.*$/ { print $0; }'
 				)"
 
@@ -1511,7 +1527,8 @@ function test_custom_ssh_configuration ()
 
 			it "Logs a redacted value."
 				user_password="$(
-					docker logs ssh.pool-1.1.1 \
+					docker logs \
+						ssh.pool-1.1.1 \
 					| awk '/^password :.*$/ { print $0; }'
 				)"
 
@@ -1579,7 +1596,8 @@ function test_custom_ssh_configuration ()
 
 			it "Logs a redacted value."
 				user_password="$(
-					docker logs ssh.pool-1.1.1 \
+					docker logs \
+						ssh.pool-1.1.1 \
 					| awk '/^password :.*$/ { print $0; }'
 				)"
 
@@ -1628,7 +1646,8 @@ function test_custom_ssh_configuration ()
 
 			it "Logs the setting value."
 				timezone="$(
-					docker logs ssh.pool-1.1.1 \
+					docker logs \
+						ssh.pool-1.1.1 \
 					| awk '/^timezone :.*$/ { print $0; }'
 				)"
 
@@ -1654,8 +1673,9 @@ function test_custom_ssh_configuration ()
 			sleep ${STARTUP_TIME}
 
 			it "Can disable sshd-bootstrap."
-				docker logs ssh.pool-1.1.1 \
-					| grep -qE 'INFO success: sshd-bootstrap entered RUNNING state'
+				docker logs \
+					ssh.pool-1.1.1 \
+				| grep -qE 'INFO success: sshd-bootstrap entered RUNNING state'
 
 				assert equal \
 					"${?}" \
@@ -1699,6 +1719,8 @@ function test_custom_sftp_configuration ()
 {
 	local container_port_22=""
 	local chroot_path=""
+	local password=""
+	local pwd_result=""
 	local user_shell=""
 
 	describe "Customised SFTP configuration"
@@ -1709,11 +1731,96 @@ function test_custom_sftp_configuration ()
 			exit 1" \
 			INT TERM EXIT
 
-		__terminate_container \
-			sftp.pool-1.1.1 \
-		&> /dev/null
+		describe "Configure password authentication"
+			__terminate_container \
+				sftp.pool-1.1.1 \
+			&> /dev/null
+
+			docker run \
+				--detach \
+				--name sftp.pool-1.1.1 \
+				--env "SSH_PASSWORD_AUTHENTICATION=true" \
+				--env "SSH_USER_FORCE_SFTP=true" \
+				--publish ${DOCKER_PORT_MAP_TCP_22}:22 \
+				jdeathe/centos-ssh:latest \
+			&> /dev/null
+
+			container_port_22="$(
+				__get_container_port \
+					sftp.pool-1.1.1 \
+					22/tcp
+			)"
+
+			if ! __is_container_ready \
+				sftp.pool-1.1.1 \
+				${STARTUP_TIME} \
+				"/usr/sbin/sshd " \
+				"grep \
+					'^Server listening on 0\.0\.0\.0 port 22\.' \
+					/var/log/secure"
+			then
+				exit 1
+			fi
+
+			it "Can connect using password authentication."
+				password="$(
+					docker logs \
+						sftp.pool-1.1.1 \
+					| awk '/^password :.*$/ { print $3; }'
+				)"
+
+				pwd_result="$(
+					expect test/sftp-password-auth.exp \
+						"${password}" \
+						app-admin \
+						"${DOCKER_HOSTNAME}" \
+						"${container_port_22}" \
+						"pwd" \
+					| tail -n 2 \
+					| head -n 1 \
+					| awk -F": " '{ print $NF; }' \
+					| tr -d '\r'
+				)"
+
+				assert equal \
+					"${pwd_result}" \
+					"/"
+			end
+
+			it "Removes insecure public key."
+				docker exec \
+					sftp.pool-1.1.1 \
+					bash -c \
+						"if [[ ! -s /home/app-admin/.ssh/authorized_keys ]]; then \
+							exit 0; \
+						else \
+							exit 1; \
+						fi" \
+				&> /dev/null
+
+				assert equal \
+					"${?}" \
+					0
+			end
+
+			it "Logs the setting value."
+				password_authentication="$(
+					docker logs \
+						sftp.pool-1.1.1 \
+					| awk '/^password authentication :.*$/ { print $0; }'
+				)"
+
+				assert equal \
+					"${password_authentication/password authentication : /}" \
+					"yes"
+			end
+		end
 
 		describe "Configure a ChrootDirectory"
+			__terminate_container \
+				sftp.pool-1.1.1 \
+			&> /dev/null
+
 			docker run \
 				--detach \
 				--name sftp.pool-1.1.1 \
@@ -1762,7 +1869,8 @@ function test_custom_sftp_configuration ()
 
 			it "Logs the setting value."
 				chroot_path="$(
-					docker logs sftp.pool-1.1.1 \
+					docker logs \
+						sftp.pool-1.1.1 \
 					| awk '/^chroot path :.*$/ { print $0; }'
 				)"
 
@@ -1821,18 +1929,23 @@ function test_custom_sftp_configuration ()
 
 				it "Will not set the key."
 					docker exec sftp.pool-1.1.1 \
-						ls \
-						/home/app-admin/.ssh/id_rsa \
-						2> /dev/null
+						bash -c \
+							"if [[ ! -s /home/app-admin/.ssh/id_rsa ]]; then \
+								exit 0; \
+							else \
+								exit 1; \
+							fi" \
+					&> /dev/null
 
-					assert unequal \
+					assert equal \
 						"${?}" \
 						"0"
 				end
 
 				it "Logs N/A key signature."
 					user_key_signature="$(
-						docker logs sftp.pool-1.1.1 \
+						docker logs \
+							sftp.pool-1.1.1 \
 						| sed -n -e '/^rsa private key fingerprint :$/{ n; p; }' \
 						| awk '{ print $1; }'
 					)"